Refs #402: Sync'ed external id types to all be ints.
git-svn-id: https://motech.rcg.usm.maine.edu/svn/motech/trunk@519 6f28304a-7744-dd4f-b2a8-8a00c5622ee2

+++ motech-server-core/src/main/java/org/motechproject/server/model/Facility.java
@@ -10,7 +10,7 @@ public class Facility implements Serializable {
 	private static final long serialVersionUID = -791545747208269078L;
 
 	private Long id;
-	private Long facilityId;
+	private Integer facilityId;
 	private Location location;
 	private String phoneNumber;
 	private Set<Community> communities;
@@ -23,11 +23,11 @@ public class Facility implements Serializable {
 		this.id = id;
 	}
 
-	public Long getFacilityId() {
+	public Integer getFacilityId() {
 		return facilityId;
 	}
 
-	public void setFacilityId(Long facilityId) {
+	public void setFacilityId(Integer facilityId) {
 		this.facilityId = facilityId;
 	}
